requirements.txt
fastapi==0.70.1
pydantic==1.8.2
pyodbc==5.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
uvicorn==0.24.0.post1
